requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.103.0
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
